move optimization info to utils
